Fixed bugs in 1D and 2D modesolver related to boundary conditions It turns out that with first order finite differences, the boundary conditions of the simulation region need to be handled with some care. In particular, it turns out to be quite important to enforce the requirement that the parallel components of E be continuous across the boundary and the perpendicular components of H be continuous.
Previously, these boundary conditions were not strictly enforced. As a result, the mode solvers would generate very non-physical results for very low index constrast systems. This has now been fixed in both the 1D and 2D mode solvers, and they should now be pretty reliable.

Note: As of now, the divergence equations for B and E are not enforced within the simulation, but only at the boundaries. This makes the mode solver susceptible to spurious solutions. In most cases, this is not likely to cause problems, however a solution should be investigated in the future.